Do get in touch section and footer
Used paddings and some flex for the get in touch section to make them
look nice and consistent with the pages layout and coloring
enjoy. Oh and the get in touch isn't working for now, I'll update
that when i learn some more javascript and css animations, I will
like to implement a pop-up that carries my contact info. So stay
tuned for that.